refactor(model): decompose main class
